+++ app/components/index.ts
@@ -8,6 +8,7 @@ export * from './CatchContainer'
 export * from './Checkbox'
 export * from './Collapsible'
 export * from './ConfirmDelete'
+export * from './CopyClick'
 export * from './CreateNewButton'
 export * from './Divider'
 export * from './Drawer'
