+++ app/routes/$bandId/band/newMember.tsx
@@ -1,14 +1,10 @@
-import { faCopy } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import type { ActionArgs, LoaderArgs } from "@remix-run/server-runtime";
 import { json, redirect } from "@remix-run/node";
-import { Button, CatchContainer, ErrorContainer, FlexList, Label } from "~/components";
+import { Button, CatchContainer, CopyClick, ErrorContainer, FlexList, Label } from "~/components";
 import { requireAdminMember } from "~/session.server";
 import invariant from "tiny-invariant";
 import { getBand, updateBandCode } from "~/models/band.server";
 import { useFetcher, useLoaderData } from "@remix-run/react";
-import { useEffect, useState } from "react";
-import { AnimatePresence, motion } from "framer-motion";
 
 export async function loader({ request, params }: LoaderArgs) {
   const bandId = params.bandId
@@ -36,21 +32,7 @@ export async function action({ request, params }: ActionArgs) {
 
 export default function NewMember() {
   const { band } = useLoaderData<typeof loader>()
-  const [showSuccess, setShowSuccess] = useState(false)
   const fetcher = useFetcher()
-
-  const handleCopy = () => {
-    navigator.clipboard.writeText(band.code).then(() => setShowSuccess(true))
-  }
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      if (showSuccess) {
-        setShowSuccess(false)
-      }
-    }, 1000)
-    return () => clearTimeout(timer)
-  }, [showSuccess])
 
   return (
     <FlexList pad={4}>
@@ -61,25 +43,7 @@ export default function NewMember() {
 
       <FlexList gap={0}>
         <Label>Band code</Label>
-        <div className="relative">
-          <button onClick={handleCopy} className=" rounded border flex justify-between items-center p-2 w-full">
-            <span className="font-bold">{band.code}</span>
-            <FontAwesomeIcon icon={faCopy} />
-          </button>
-          <AnimatePresence>
-            {showSuccess ? (
-              <motion.span
-                initial={{ opacity: 0, y: -10 }}
-                animate={{ opacity: 1, y: 0 }}
-                exit={{ opacity: 0, y: -10 }}
-                transition={{ opacity: .2 }}
-                className="absolute top-full right-0 text-primary bg-white p-2 rounded"
-              >
-                Band code copied!
-              </motion.span>
-            ) : null}
-          </AnimatePresence>
-        </div>
+        <CopyClick textToCopy={band.code} copyMessage={band.code} successMessage="Band code copied!" />
       </FlexList>
 
       <fetcher.Form method="put" action=".">
