+++ app/routes/$bandId/setlists.tsx
@@ -136,7 +136,7 @@ export default function SetlistsRoute() {
             {hasSetlists ? (
               <FlexList pad={4} gap={2}>
                 {setlists.map(setlist => (
-                  <SetlistLink key={setlist.id} setlist={setlist} />
+                  <SetlistLink key={setlist.id} setlist={setlist} publicRemark="You can remove the public link by clicking on this setlist's settings." />
                 ))}
               </FlexList>
             ) : (
